+++ mapper/src/main/java/org/grnet/status/mappers/IncidentMapper.java
@@ -2,7 +2,9 @@ package org.grnet.status.mappers;
 
 import org.grnet.status.dtos.incident.IncidentRequestDto;
 import org.grnet.status.dtos.incident.IncidentResponseDto;
+import org.grnet.status.dtos.incident.ServiceDto;
 import org.grnet.status.entities.Incident;
+import org.grnet.status.entities.IncidentAffectedService;
 import org.mapstruct.IterableMapping;
 import org.mapstruct.Mapper;
 import org.mapstruct.Mapping;
@@ -18,8 +20,6 @@ public interface IncidentMapper {
 
     IncidentMapper INSTANCE = Mappers.getMapper(IncidentMapper.class);
 
-    @Mapping(source = "service.id", target = "serviceId")
-    @Mapping(source = "service.name", target = "serviceName")
     @Mapping(target = "id", ignore = true)
     @Mapping(target = "tenant", ignore = true)
     @Mapping(target = "incidentNumber", ignore = true)
@@ -31,12 +31,19 @@ public interface IncidentMapper {
     @Mapping(target = "comments", ignore = true)
     Incident incidentRequestToEntity(IncidentRequestDto request);
 
+    @Mapping(source = "id", target = "serviceId")
+    @Mapping(source = "name", target = "serviceName")
+    @Mapping(target = "id", ignore = true)
+    @Mapping(target = "incident", ignore = true)
+    IncidentAffectedService serviceGroupDtoToEntity(ServiceDto service);
+
     @IterableMapping(qualifiedByName = "incidentMap")
     List<IncidentResponseDto> incidentsToDtos(List<Incident> incidents);
 
     @Named("incidentMap")
-    @Mapping(source = "serviceId", target = "service.id")
-    @Mapping(source = "serviceName", target = "service.name")
     IncidentResponseDto incidentToResponseDto(Incident incident);
 
+    @Mapping(source = "serviceId", target = "id")
+    @Mapping(source = "serviceName", target = "name")
+    ServiceDto incidentServiceGroupToDto(IncidentAffectedService service);
 }